requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
openai==1.50.0
google-generativeai==0.8.0
python-dotenv==1.0.1
websockets==13.0
httpx==0.27.0
